Make log repo more optional

+++ core/src/main/java/com/example/rules/core/repository/RuleLogRepository.java
@@ -10,7 +10,7 @@ import java.util.stream.Stream;
 @Repository
 public interface RuleLogRepository extends CrudRepository<RuleLog, Long> {
 
-    Stream<RuleLog> findByRequestClassAndRequestHashOrderByCreateTimeDesc(String requestClass, long requestHash);
+    Stream<RuleLog> findByRequestClassAndRequestHashOrderByCreateTimeDesc(String requestClass, int requestHash);
 
     void deleteByCreateTimeBefore(LocalDateTime createTime);
 }
